Separated zk ops
