+++ src/app_legacy.rs
@@ -11,31 +11,11 @@ use ratatui::{
 use std::{cmp::min, mem, net::IpAddr, vec};
 use zookeeper_async::{Acl, Stat};
 
-use crate::{app::{App, BASE_RESOURCE, CONFIRMATION_STRING}, node_data::NodeData, tab::Tab};
-
-// const BASE_RESOURCE: &str = "/";
-// const CONFIRMATION_STRING: &str = "DELETE";
-// #[derive(Default)]
-// pub struct App {
-//     pub state: AppState,
-//     pub connection: Option<Connection>,
-//     pub zk: Option<zookeeper_async::ZooKeeper>,
-//     pub connection_input: String,
-//     pub curr_tab: usize,
-//     pub tabs: Vec<Tab>,
-
-//     // -//- Moved to Tab struct
-//     pub tab_data: Vec<String>,
-//     pub list_state: ListState,
-//     pub curr_resource: Option<String>, // selected node for current nest level
-//     pub prev_resources: Vec<String>,   // prev resources: e.g. /zookeeper/config
-//     pub current_node_stat: Option<Stat>,
-//     pub message: String,
-//     pub node_data: NodeData,
-//     pub node_path_buf: String,
-//     pub node_data_buf: String,
-//     pub input_buf: String,
-// }
+use crate::{
+    app::{App, BASE_RESOURCE, CONFIRMATION_STRING},
+    node_data::NodeData,
+    tab::Tab,
+};
 
 impl App {
     pub fn selected_resource(&self) -> Option<String> {
@@ -48,20 +28,6 @@ impl App {
 
     pub fn set_current_tab_path(&mut self, path: Option<String>) {
         self.curr_resource = path;
-    }
-
-    pub(crate) async fn store_node_stat(&mut self) {
-        let full_path = self.full_resource_path();
-        let _ = self
-            .zk
-            .as_ref()
-            .unwrap()
-            .exists(&full_path, false)
-            .and_then(|stat| async {
-                self.current_node_stat = stat;
-                Ok(())
-            })
-            .await;
     }
 
     pub(crate) fn full_resource_path(&self) -> String {
@@ -77,24 +43,6 @@ impl App {
     pub(crate) fn is_full_resources_path_empty(&self) -> bool {
         self.prev_resources.is_empty() && self.curr_resource.is_none()
     }
-    pub(crate) async fn store_children(&mut self) {
-        {
-            self.clear_message();
-            self.append_message(format!("Full path : {}\n", self.full_resource_path()));
-            self.append_message(format!("Prev path : {:?}\n", self.prev_resources));
-        }
-        let Some(ref zk) = self.zk else {
-            return;
-        };
-
-        let children = zk
-            .get_children(&self.full_resource_path(), false)
-            .await
-            .ok();
-        if let Some(ch) = children {
-            self.tab_data = ch;
-        }
-    }
 
     pub(crate) fn set_message(&mut self, msg: String) {
         self.message = msg;
@@ -106,73 +54,6 @@ impl App {
 
     pub(crate) fn clear_message(&mut self) {
         self.message.clear();
-    }
-
-    pub(crate) async fn store_node_data(&mut self) {
-        let Some(ref zk) = self.zk else {
-            return;
-        };
-
-        let _ = zk
-            .get_data(&self.full_resource_path(), false)
-            .and_then(|(data, _)| async {
-                self.node_data = NodeData::Raw(data);
-                Ok(())
-            })
-            .await;
-    }
-
-    pub(crate) async fn create_node(&mut self) {
-        let Some(ref zk) = self.zk else {
-            "Failed to get zookeeper client".clone_into(&mut self.message);
-            return;
-        };
-
-        let res = zk
-            .create(
-                &self.node_path_buf,
-                self.node_data_buf.clone().into_bytes(),
-                Acl::open_unsafe().clone(),
-                zookeeper_async::CreateMode::Persistent,
-            )
-            .await;
-        match res {
-            Ok(created_path) => self.message = format!("Node {created_path} created successfully"),
-            Err(e) => self.message = format!("Node creation failed : {e}"),
-        }
-    }
-
-    pub(crate) async fn set_data(&mut self) {
-        let Some(ref zk) = self.zk else {
-            "Failed to get zookeeper client".clone_into(&mut self.message);
-            return;
-        };
-
-        let data = mem::take(&mut self.node_data_buf).into_bytes();
-        let res = zk.set_data(&self.full_resource_path(), data, None).await;
-        match res {
-            Ok(_) => {
-                self.message = format!(
-                    "Node {} data successfully updated",
-                    self.full_resource_path()
-                )
-            }
-            Err(e) => self.message = format!("Node data update failed : {e}"),
-        }
-    }
-
-    pub(crate) async fn delete_node(&mut self) {
-        let Some(ref zk) = self.zk else {
-            "Failed to get zookeeper client".clone_into(&mut self.message);
-            return;
-        };
-        let res = zk.delete(&self.full_resource_path(), None).await;
-        match res {
-            Ok(_) => {
-                self.message = format!("Node {} successfully deleted", self.full_resource_path())
-            }
-            Err(e) => self.message = format!("Delete node failed : {e}"),
-        }
     }
 
     pub(crate) fn is_deletion_confirmed(&mut self) -> bool {
